enables mod1 fault counting for single fault competencies
