+++ src/providers/fault-count/cat-c/fault-count.cat-c.ts
@@ -7,7 +7,7 @@ import { CatC1EUniqueTypes } from '@dvsa/mes-test-schema/categories/C1E';
 import { sumManoeuvreFaults } from '../../../shared/helpers/faults';
 import { CompetencyOutcome } from '../../../shared/models/competency-outcome';
 import { VehicleChecksScore } from '../../../shared/models/vehicle-checks-score.model';
-import { getCompetencyFaults } from '../../../shared/helpers/competency';
+import { getCompetencyFaults } from '../../../shared/helpers/get-competency-faults';
 
 export class FaultCountCHelper {
