+++ src/providers/fault-summary/cat-a-mod1/__tests__/cat-AM1-test-data-mock.ts
@@ -12,10 +12,10 @@ export const catAM1TestDataStateObject: TestData = {
     controlledStop: CompetencyOutcome.S,
   },
   emergencyStop: {
-    outcome: CompetencyOutcome.D,
+    outcome: CompetencyOutcome.DF,
   },
   avoidance: {
-    outcome: CompetencyOutcome.S,
+    outcome: CompetencyOutcome.DF,
   },
   drivingFaults: {
     precautions: 2,
