+++ src/providers/fault-summary/cat-c/fault-summary.cat-c.ts
@@ -5,7 +5,7 @@ import { CompetencyDisplayName } from '../../../shared/models/competency-display
 import { CompetencyOutcome } from '../../../shared/models/competency-outcome';
 import { ManoeuvreTypes } from '../../../modules/tests/test-data/test-data.constants';
 import { TestCategory } from '@dvsa/mes-test-schema/category-definitions/common/test-category';
-import { getCompetencyFaults, getCompetencyComment } from '../../../shared/helpers/competency';
+import { getCompetencyFaults, getCompetencyComment } from '../../../shared/helpers/get-competency-faults';
 import { CatCUniqueTypes } from '@dvsa/mes-test-schema/categories/C';
 import {
   manoeuvreCompetencyLabelsCatC,
